Small refactor of image attachments…
Instead of CCoreTextRenderer handling the image attachment delegate, a new class CCoreTextAttachment exists to act as a bridge between CMarkupValueTransformer and image rendering. You now create an attachment object with the parameters you need (including a block that will render the attachment) and embed it in the NSAttributedString. When CCoreTextRenderer encounters this attachment it calls the attachment's renderer block.

This means that CCoreTextRenderer no longer needs to "normalise" text and no longer needs to cache that information.

More work upfront (where it belongs). Less at render time.